fix: repair lib imports after netlify function directory move (#96)
